page-2-grpc/requirements.txt
Flask==3.0.0
grpcio==1.68.1
grpcio-tools==1.68.1